Adapting to VTKs double changes

+++ ParaView/ParaView/vtkPVRenderView.h
@@ -143,8 +143,11 @@ public:
 
   // Description:
   // Change the background color.
-  void SetBackgroundColor(float r, float g, float b);
-  virtual void SetBackgroundColor(float *c) {this->SetBackgroundColor(c[0],c[1],c[2]);}
+  void SetBackgroundColor(double r, double g, double b);
+  virtual void SetBackgroundColor(double *c) 
+    {
+      this->SetBackgroundColor(c[0],c[1],c[2]);
+    }
 
   // Description:
   // Close the view - called from the vtkkwwindow. This default method
